Main text book to study

J2L: Julia as a second language, Erik Engheim, 2023

JDP: Hands on design patterns and best practices with julia, Tom Kwong, 2020

Study Schedule (2023)

Date	Range	Presenter

Aug 2	Setup Julia Repo in GitHub with Actions	Norel

Aug 10	Setup Julia Repo in GitHub with Actions, Why Julia?	Norel

Aug 24	Why Julia?	Kjeong

Sep 15	Organizing and modularizing your code, Modules, Packages and Data Type Concepts	Norel

Sep 24	Julia as a calculator	Kjeong

Oct 6	Control flow	Norel

Oct 13	Julia as a spreadsheet	kjeong

Oct 20	Working with text	Norel

Oct 27	Storing data in dictionaries	kjeong

Nov 3	Understanding types, Modules, Packages and Data Type Concepts	Norel

Documentation for JuliaStudy.

JuliaStudy.JuliaPatterns.DomainSpecificLanguagePattern.LModel

JuliaStudy.JuliaPatterns.DomainSpecificLanguagePattern.LState

Base.:==

Base.show

JuliaStudy.Julia2ndLang.ToyGeometry.cone_value

JuliaStudy.Julia2ndLang.ToyGeometry.cosine

JuliaStudy.Julia2ndLang.ToyGeometry.cylinder_volume

JuliaStudy.Julia2ndLang.ToyGeometry.fac

JuliaStudy.Julia2ndLang.ToyGeometry.sine

JuliaStudy.Julia2ndLang.ToyGeometry.sphere_volume

JuliaStudy.JuliaPatterns.Calculator.interest

JuliaStudy.JuliaPatterns.Calculator.rate

JuliaStudy.JuliaPatterns.DataTypeConcepts.genSubStr

JuliaStudy.JuliaPatterns.DataTypeConcepts.location

JuliaStudy.JuliaPatterns.DataTypeConcepts.subtypeTreeStr

JuliaStudy.JuliaPatterns.DataTypeConcepts.subtypetree

JuliaStudy.JuliaPatterns.DomainSpecificLanguagePattern.add_rule!

JuliaStudy.JuliaPatterns.DomainSpecificLanguagePattern.result

JuliaStudy.JuliaPatterns.SharedArrayPattern.std_by_security

JuliaStudy.JuliaPatterns.SpaceWar.explode

JuliaStudy.JuliaPatterns.SpaceWar.overlap

JuliaStudy.JuliaPatterns.SpaceWar.tow

JuliaStudy.JuliaPatterns.Vehicle.engage_wheels!

JuliaStudy.JuliaPatterns.Vehicle.land2!

JuliaStudy.JuliaPatterns.Vehicle.move!

JuliaStudy.JuliaPatterns.Vehicle.position

JuliaStudy.JuliaPatterns.Vehicle.power_off!

JuliaStudy.JuliaPatterns.Vehicle.power_on!

JuliaStudy.JuliaPatterns.Vehicle.travel_path

JuliaStudy.JuliaPatterns.Vehicle.turn!

Base.:== — Method

Base.:(==)(a::Archer, b::Archer)

See https://stackoverflow.com/questions/70362843/how-to-create-equality-test-case-for-custom-structures-in-julia

JuliaStudy.Julia2ndLang.ToyGeometry.cone_value — Method

cone_value(r, h)

Calculate the volume of a cone with radius r and height h.

JuliaStudy.Julia2ndLang.ToyGeometry.cosine — Method

cosine(x::Number)

Calculate the cosine of an angle x given in radians

JuliaStudy.Julia2ndLang.ToyGeometry.cylinder_volume — Method

cylinder_volume(r, h)

Calculate the volume of a cylinder with radius r and height h.

JuliaStudy.Julia2ndLang.ToyGeometry.fac — Method

fac(n)

Get factorial of n.

JuliaStudy.Julia2ndLang.ToyGeometry.sine — Method

sine(x)

Calculate the sine of an angle x given in radians.

JuliaStudy.Julia2ndLang.ToyGeometry.sphere_volume — Method

sphere_volume(r)

Calculate the volume of a sphere with radius r

JuliaStudy.JuliaPatterns.Calculator.interest — Method

interest(amount, rate)

Calculate interest from an amount and interest rate of rate.

JuliaStudy.JuliaPatterns.Calculator.rate — Method

rate(amount, interest)

Calculate interest rate based on an amount and interest.

JuliaStudy.JuliaPatterns.DataTypeConcepts.genSubStr — Method

genSubStr(inType, indent::Int, level::Int, inStr::String) ::String

Generate substring for recursive call from subtypeTreeStr.

JuliaStudy.JuliaPatterns.DataTypeConcepts.location — Method

location(p::Property)

Returns the location of the property as a tuple of (latitude, longitude).

JuliaStudy.JuliaPatterns.DataTypeConcepts.subtypeTreeStr — Function

subtypeTreeStr(inType = nothing, indent::Int = 4) ::String

Return all sub types with tree stucture.

This function is improved version of subtypetree for easy testing.

JuliaStudy.JuliaPatterns.DataTypeConcepts.subtypetree — Function

subtypetree(roottype, level = 1, indent = 4)

Display the entire type hierarchy starting from the specified roottype

JuliaStudy.JuliaPatterns.SharedArrayPattern.std_by_security — Method

std_by_security(valuation)

Find standard deviation of each attribute for each security

JuliaStudy.JuliaPatterns.SpaceWar.explode — Method

explode(things::AbstractVector{Any})

explode an array of objects

JuliaStudy.JuliaPatterns.SpaceWar.overlap — Method

overlap(A::Rectangle, B::Rectangle)

check if two rectangles (A & B) overlap

JuliaStudy.JuliaPatterns.SpaceWar.tow — Method

tow(A::Spaceship, B::T) where {T <: Thing}

equivalent of parametric type

JuliaStudy.JuliaPatterns.Vehicle.engage_wheels! — Method

engage_wheels(args...)

engage wheels for landing. Optional

JuliaStudy.JuliaPatterns.Vehicle.land2! — Method

land2!(vehicle)

Landing (using trait)

JuliaStudy.JuliaPatterns.Vehicle.move! — Function

move!

move the vehicle by the specified distance

JuliaStudy.JuliaPatterns.Vehicle.position — Function

position

returns the (x, y) position of the vehicle

JuliaStudy.JuliaPatterns.Vehicle.power_off! — Function

power_off!

turn off the vehicle's engine

JuliaStudy.JuliaPatterns.Vehicle.power_on! — Function

power_on!

turn on the vehicle's engine

JuliaStudy.JuliaPatterns.Vehicle.travel_path — Method

travel_path(position, destination)

Returns a travel plan from current position to destination

JuliaStudy.JuliaPatterns.Vehicle.turn! — Function

turn!

steer vehicle to the specified direction

JuliaStudy.JuliaPatterns.DomainSpecificLanguagePattern.LModel — Method

Create a L-system model.

JuliaStudy.JuliaPatterns.DomainSpecificLanguagePattern.LState — Method

Create a L-system state from a model.

Base.show — Method

Display model nicely.

JuliaStudy.JuliaPatterns.DomainSpecificLanguagePattern.add_rule! — Method

Add rule to a model.

JuliaStudy.JuliaPatterns.DomainSpecificLanguagePattern.result — Method

Compact the result suitable for display